// src/enums/skills.ts
export enum Skills {
  Catfall = "Catfall",
  Dodge = "Dodge",
  JumpBack = "Jump Back",
  Leap = "Leap",
  QuickWitted = "Quick Witted",
  Sprint = "Sprint",
  CombatMaster = "Combat Master",
  Disarm = "Disarm",
  Feint = "Feint",
  Deflect = "Deflect",
  CounterAttack = "Counter Attack",
  StepAside = "Step Aside",
  Ambush = "Ambush",
  Dive = "Dive",
  EscapeArtist = "Escape Artist",
  Evade = "Evade",
  Infiltration = "Infiltration",
  SneakUp = "Sneak Up",
  BodySlam = "Body Slam",
  BulgingBiceps = "Bulging Biceps",
  HardAsNails = "Hard as Nails",
  Juggernaut = "Juggernaut",
  HurlOpponent = "Hurl Opponent",
  IronJaw = "Iron Jaw",
  BerserkCharge = "Berserk Charge",
  Impetuous = "Impetuous",
  IronWill = "Iron Will",
  KillerReputation = "Killer Reputation",
  NervesOfSteel = "Nerves of Steel",
  TrueGrit = "True Grit",
  CrackShot = "Crack Shot",
  FastShot = "Fast Shot",
  Gunfighter = "Gunfighter",
  HipShooting = "Hip Shooting",
  Marksman = "Marksman",
  RapidFire = "Rapid Fire",
  Armourer = "Armourer",
  Fixer = "Fixer",
  Inventor = "Inventor",
  Medic = "Medic",
  Specialist = "Specialist",
  Weaponsmith = "Weaponsmith",
}